+++ Android/UdacityExercises/Lesson03-Green-Recycler-View/T03.02-Exercise-ViewHolder/app/src/main/java/com/example/android/recyclerview/GreenAdapter.java
@@ -26,30 +26,16 @@ import android.widget.TextView;
 /**
  * We couldn't come up with a good name for this class. Then, we realized
  * that this lesson is about RecyclerView.
- *
+ * <p>
  * RecyclerView... Recycling... Saving the planet? Being green? Anyone?
  * #crickets
- *
+ * <p>
  * Avoid unnecessary garbage collection by using RecyclerView and ViewHolders.
- *
+ * <p>
  * If you don't like our puns, we named this Adapter GreenAdapter because its
  * contents are green.
  */
 public class GreenAdapter extends RecyclerView.Adapter<GreenAdapter.NumberViewHolder> {
-
-    // TODO (1) Create a layout resource in res/layout/ called number_list_item.xml
-
-    // Do steps 2 - 11 within number_list_item.xml
-    // TODO (2) Make the root layout a FrameLayout
-    // TODO (3) Make the width match_parent and the height wrap_content
-    // TODO (4) Set the padding to 16dp
-    // TODO (5) Add a TextView as the only child of the FrameLayout
-    // TODO (6) Give the TextView an ID "@+id/tv_item_number"
-    // TODO (7) Set the height and width to wrap_content
-    // TODO (8) Align the TextView to the start of the parent
-    // TODO (9) Center the TextView vertically in the layout
-    // TODO (10) Set the font family to monospace
-    // TODO (11) Set the text size to 42sp
 
     private static final String TAG = GreenAdapter.class.getSimpleName();
 
@@ -66,7 +52,6 @@ public class GreenAdapter extends RecyclerView.Adapter<GreenAdapter.NumberViewHo
     }
 
     /**
-     *
      * This gets called when each new ViewHolder is created. This happens when the RecyclerView
      * is laid out. Enough ViewHolders will be created to fill the screen and allow for scrolling.
      *
@@ -117,16 +102,17 @@ public class GreenAdapter extends RecyclerView.Adapter<GreenAdapter.NumberViewHo
         return mNumberItems;
     }
 
-    // TODO (12) Create a class called NumberViewHolder that extends RecyclerView.ViewHolder
+    class NumberViewHolder extends RecyclerView.ViewHolder {
+        TextView listItemNumberView;
 
-    // TODO (13) Within NumberViewHolder, create a TextView variable called listItemNumberView
+        NumberViewHolder(View itemView) {
+            super(itemView);
 
-    // TODO (14) Create a constructor for NumberViewHolder that accepts a View called itemView as a parameter
-    // TODO (15) Within the constructor, call super(itemView) and then find listItemNumberView by ID
+            listItemNumberView = (TextView) itemView.findViewById(R.id.tv_item_number);
+        }
 
-    // TODO (16) Within the NumberViewHolder class, create a void method called bind that accepts an int parameter called listIndex
-    // TODO (17) Within bind, set the text of listItemNumberView to the listIndex
-    // TODO (18) Be careful to get the String representation of listIndex, as using setText with an int does something different
-
+        void bind(int listIndex) {
+            listItemNumberView.setText(String.valueOf(listIndex));
+        }
     }
 }