+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 27.08.2018
 
+* [Android] RecyclerView: ViewHolder 
 * [Android] RecyclerView: Create RecyclerView
 
 #### 26.08.2018